requirements_fastapi.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
google-generativeai
python-dotenv
docx2pdf
fpdf2
python-magic-bin